pyproject.toml
[tool.poetry]
name = "video2text"
version = "0.1.0"
description = ""
authors = ["TomoyaFujita2016 <fujita.t.2016c@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
youtube_dl = "^2021.12.17"
webvtt-py = "^0.4.6"
spacy = "^3.2.1"
ginza = "^5.1.0"
ja-ginza = "^5.1.0"
ja-ginza-electra = "^5.1.0"
ginza-transformers = "^0.4.0"
spacy-transformers = "1.1.3"
tqdm = "^4.62.3"
sumy = "^0.9.0"
tinysegmenter = "^0.4"
streamlit = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
